+++ CHANGELOG.md
@@ -14,6 +14,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Do not use Homebridge UI > v5.5.0 because of break config.json
 
+## [4.9.0] - (11.12.2025)
+
+## Changes
+
+- added recording sensor
+- added streaming sensor
+- config schema updated
+- redme updated
+- cleanup
+
 ## [4.8.20] - (09.12.2025)
 
 ## Changes